Fix more stubs and string class names.
