+++ composer.json
@@ -73,8 +73,7 @@
             "TestPluginTwo\\": "tests/test_app/Plugin/TestPluginTwo/src/",
             "Company\\TestPluginThree\\": "tests/test_app/Plugin/Company/TestPluginThree/src/",
             "Company\\TestPluginThree\\Test\\": "tests/test_app/Plugin/Company/TestPluginThree/tests/",
-            "ParentPlugin\\": "tests/test_app/Plugin/ParentPlugin/src/",
-            "PluginJs\\": "tests/test_app/Plugin/PluginJs/src/"
+            "ParentPlugin\\": "tests/test_app/Plugin/ParentPlugin/src/"
         }
     },
     "replace": {
