+++ tests/test_app/TestApp/Error/TestDebugger.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Error;
+
+use Cake\Error\Debugger;
+
+class TestDebugger extends Debugger
+{
+}
