+++ tests/test_app/TestApp/Model/Entity/NumberTree.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Model\Entity;
+
+use Cake\ORM\Behavior\Translate\TranslateTrait;
+use Cake\ORM\Entity;
+
+/**
+ * Stub entity class
+ */
+class NumberTree extends Entity
+{
+    use TranslateTrait;
+}
